minor code refactoring
